Update render_graphs.py
Updated to release 04 with a couple functional changes
a) added a red line to denote the maximum number of licenses owned for each database, Edit this file and update hline in the IF statement group to match your owned licenses.
b) added a 3 month option for timeframe. Do note that this requires editing the radio button call in index.html to match the function call to render graphs
c) Be sure to change the path for your install (Mine is AIX 6.3L with GNU tools) as well, update the databases to reflect yours
d) Thanks John for writing such a great tool. To comply with the GPL V3 and share my changes with everyone I'm doing my part to give back! Thanks everyone.