Add files via upload
Adding a brief install gide, tips and tricks, FAQ and TESTED WITH document to help people trying to implement this. Anyone who gets this far and reads this please add your tips/tricks, add your name, and the daemons you've tested this on.

Thanks everyone and again thanks John for writing this great tool.